requirements.txt
asgiref>=3.7.2
certifi>=2024.2.2
chardet>=5.2.0
defusedxml>=0.7.1
dj-database-url>=2.1.0
Django==5.2.1
django-allauth>=0.61.1
django-ckeditor>=6.7.0
django-js-asset>=2.0.0
django-multiselectfield>=0.1.12
gunicorn>=21.2.0
idna>=3.6
oauthlib>=3.2.2
Pillow>=10.3.0
psycopg2>=2.9.9
psycopg2-binary>=2.9.9
python3-openid>=3.2.0
pytz>=2024.1
requests>=2.31.0
requests-oauthlib>=1.3.1
sqlparse>=0.5.0
urllib3>=2.2.1
whitenoise>=6.6.0
numpy-2.2.5 opencv-python-4.11.0.86 
torch torchvision pillow numpy
colorthief
geopy
pip install roboflow --user
pip install opencv-python
pip install opencv-python
pip install inference
pip install ultralytics opencv-python pillow
https://codepen.io/creativesite/pen/mjoyvp
https://naver.github.io/egjs-axes/docs/car360viewer
pip install jupyter pandas numpy scikit-learn matplotlib joblib xgboost catboost seaborn
pip install notebook
pip install missingno
pip install category_encoders
pip install xgboost
#Username: praphulla 
#Password: Project@123
#for admin login